doing server variables
I'm changing the scope of the variables a lil bit. Initially I was going to make a script that you could throw on any object, but the details on how to reference the variable is strange, so I'm just doing a simple global server sort of thing.

Its possible I'll do something like that in the past, but its not really needed if I'm honest - events are a lot more clean (but do require some more thought)