Variable system problems
It is working, but the problem is that since both the object and prefab of index 0 have the same variable, they only ever set the object's. I need to separate the create functions but also if it gets set by a prefab, it needs to do both but only object for the client that object comes from. (ew)